// cpo-server/src/types/ChargingStation.ts
export default interface ChargingStation extends CreatedUpdatedProps, ChargingStationAuthorizationActions {
  id?: string;
  templateHash?: string;
  templateHashCapabilities?: string;
  templateHashTechnical?: string;
  templateHashOcppStandard?: string;
  templateHashOcppVendor?: string;
  issuer: boolean;
  public: boolean;
  siteAreaID?: string;
  siteID?: string;
  companyID?: string;
  chargePointSerialNumber: string;
  chargePointModel: string;
  chargeBoxSerialNumber: string;
  chargePointVendor: string;
  iccid: string;
  imsi: string;
  meterType: string;
  firmwareVersion: string;
  firmwareUpdateStatus?: OCPPFirmwareStatus;
  meterSerialNumber: string;
  endpoint: string;
  ocppVersion: OCPPVersion;
  ocppProtocol: OCPPProtocol;
  cloudHostIP?: string;
  cloudHostName?: string;
  lastSeen: Date;
  deleted: boolean;
  inactive: boolean;
  tokenID: string;
  forceInactive: boolean;
  manualConfiguration?: boolean;
  lastReboot: Date;
  chargingStationURL: string;
  maximumPower: number;
  masterSlave?: boolean;
  voltage: Voltage;
  excludeFromSmartCharging?: boolean;
  powerLimitUnit: ChargingRateUnitType;
  coordinates: number[];
  chargePoints: ChargePoint[];
  connectors: Connector[];
  backupConnectors: Connector[];
  remoteAuthorizations: RemoteAuthorization[];
  currentIPAddress?: string | string[];
  siteArea?: SiteArea;
  site?: Site;
  capabilities?: ChargingStationCapabilities;
  ocppStandardParameters?: KeyValue[];
  ocppVendorParameters?: KeyValue[];
  distanceMeters?: number;
  ocpiData?: ChargingStationOcpiData;
  oicpData?: ChargingStationOicpData;
  tariffID?: string;
  tariffIDs?: string[];
  evse_id?: string;
  certificateIDs?: string[];
}

export interface ChargingStationOcpiData {
  evses?: OCPIEvse[];
}

export interface ChargingStationOicpData {
  evses?: OICPEvseDataRecord[];
}
// ...
export interface Connector extends ConnectorAuthorizationActions {
  id?: string;
  connectorId: number;
  currentInstantWatts?: number;
  currentStateOfCharge?: number;
  currentTotalConsumptionWh?: number;
  currentTotalInactivitySecs?: number;
  currentInactivityStatus?: InactivityStatus;
  currentTransactionID?: number;
  currentTransactionDate?: Date;
  currentTagID?: string;
  currentUserID?: string;
  status: ChargePointStatus;
  errorCode?: string;
  info?: string;
  vendorErrorCode?: string;
  power?: number;
  type?: ConnectorType;
  voltage?: Voltage;
  amperage?: number;
  amperageLimit?: number;
  user?: User;
  statusLastChangedOn?: Date;
  numberOfConnectedPhase?: number;
  currentType?: CurrentType;
  chargePointID?: number;
  phaseAssignmentToGrid?: PhaseAssignmentToGrid;
  tariffID?: string;
  tariffIDs?: string[];
  ocpiConnecterId?: string;
  isPrivate?:boolean;
  ownerIds?:string[];
  certificateIDs?: string[];
}

export interface PhaseAssignmentToGrid {
  csPhaseL1: OCPPPhase.L1 | OCPPPhase.L2 | OCPPPhase.L3;
  csPhaseL2: OCPPPhase.L1 | OCPPPhase.L2 | OCPPPhase.L3;
  csPhaseL3: OCPPPhase.L1 | OCPPPhase.L2 | OCPPPhase.L3;
}

export interface RemoteAuthorization {
  id: string;
  connectorId: number;
  tagId: string;
  timestamp: Date;
  oicpIdentification?: OICPIdentification;
}

export interface ConnectorCurrentLimit {
  limitAmps: number;
  limitWatts: number;
  limitSource: ConnectorCurrentLimitSource;
}

export enum SiteAreaLimitSource {
  CHARGING_STATIONS = 'CS',
  SITE_AREA = 'SA',
}

export enum ConnectorCurrentLimitSource {
  CHARGING_PROFILE = 'CP',
  STATIC_LIMITATION = 'SL',
  CONNECTOR = 'CO'
}

export enum CurrentType {
  AC = 'AC',
  DC = 'DC'
}

export interface ChargePoint {
  chargePointID: number;
  currentType: CurrentType;
  voltage: Voltage;
  amperage: number;
  numberOfConnectedPhase: number;
  cannotChargeInParallel: boolean;
  sharePowerToAllConnectors: boolean;
  excludeFromPowerLimitation: boolean;
  ocppParamForPowerLimitation: string;
  power: number;
  efficiency: number;
  connectorIDs: number[];
}

export enum Voltage {
  VOLTAGE_400 = 400,
  VOLTAGE_230 = 230,
  VOLTAGE_110 = 110,
}
// ...
export enum ConnectorType {
  TYPE_2 = 'T2',
  COMBO_CCS = 'CCS',
  CHADEMO = 'C',
  TYPE_1 = 'T1',
  TYPE_3C = 'T3C',
  TYPE_1_CCS = 'T1CCS',
  DOMESTIC = 'D',
  UNKNOWN = 'U',
}

export interface ChargingStationCapabilities {
  supportStaticLimitation: boolean;
  supportChargingProfiles: boolean;
  supportCreditCard: boolean;
  supportRemoteStartStopTransaction: boolean;
  supportUnlockConnector: boolean;
  supportReservation: boolean;
  supportRFIDCard: boolean;
  supportFirmwareUpgrade?: boolean;
  supportSlave?: boolean;
  ISO15118PnCEnabled?: boolean;
}