// cpo-server/src/types/Pricing.ts
export enum PricingEntity {
  TENANT = 'Tenant',
  COMPANY = 'Company',
  SITE = 'Site',
  SITE_AREA = 'SiteArea',
  CHARGING_STATION = 'ChargingStation',
  // USER = 'User'
}
// ...
interface PricingDefinitionInternal {
  id?: string;
  entityID?: string; // id of the entity the pricing definition belongs to
  entityType?: PricingEntity; // Type of the entity this model belongs to
  name: string; // Short marketing name - e.g.: BLUE Tariff,
  description: string; // A long description to explain it, e.g.: Time-based pricing for low charging stations
  staticRestrictions?: PricingStaticRestriction;
  restrictions?: PricingRestriction;
  dimensions: PricingDimensions
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface ResolvedPricingDefinition extends PricingDefinitionInternal {
}

export default interface PricingDefinition extends PricingDefinitionInternal, CreatedUpdatedProps, AuthorizationActions {
  id: string;
  entityID: string;
  entityType: PricingEntity;
  siteID: string;
  roaming?:boolean;
  ocpiData?:PricingOcpiData;
}

export interface PricingDimensions {
  flatFee?: PricingDimension; // Flat fee, no unit
  energy?: PricingDimension; // Defined in kWh, step_size multiplier: 1 Wh
  chargingTime?: PricingDimension; // Time charging: defined in hours, step_size multiplier: 1 second
  parkingTime?: PricingDimension; // Time not charging: defined in hours, step_size multiplier: 1 second
}
// ...
export interface PricingStaticRestriction {
  validFrom?: Date;
  validTo?: Date;
  connectorType?: ConnectorType; // Connector types allowed to use this tariff
  connectorPowerkW?: number;
}

export interface PricingRestriction {
  daysOfWeek?: DayOfWeek[], // Which day(s) of the week this tariff is valid
  timeFrom?: string, // Valid from this time of the day
  timeTo?: string, // Valid until this time of the day
  minEnergyKWh?: number; // Minimum used energy in kWh, for example 20, valid from this amount of energy is used
  maxEnergyKWh?: number; // Maximum used energy in kWh, for example 50, valid until this amount of energy is used
  minDurationSecs?: number; // Minimum duration in seconds, valid for a duration from x seconds
  maxDurationSecs?: number; // Maximum duration in seconds, valid for a duration up to x seconds
}
// ...
export interface PricingOcpiData {
  tariff?: OCPITariff;
  tariffCheckedOn?: Date;
  tariffUpdatedOn?: Date;
  inEmsp?: boolean;
}